unify link on nightly and releases page
