+++ apps/web/src/lib/components/marketing/ReleaseCard.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import osIcons from '$lib/data/os-icons.json';
+	import ReleaseDownloadLinks from '$lib/components/marketing/ReleaseDownloadLinks.svelte';
 	import Markdown from 'svelte-exmarkdown';
 	import type { Release } from '$lib/types/releases';
 
@@ -13,49 +13,6 @@
 
 	// State for toggling download links visibility
 	let downloadLinksVisible = $state(false);
-
-	// Group builds by OS for easier display
-	const groupedBuilds = $derived.by(() => {
-		const builds =
-			release.builds?.reduce(
-				(acc, build) => {
-					if (!acc[build.os]) acc[build.os] = [];
-					acc[build.os].push(build);
-					return acc;
-				},
-				{} as Record<string, typeof release.builds>
-			) || {};
-
-		// Sort OS entries: macOS (darwin) first, then Windows, then Linux
-		const osOrder = ['darwin', 'windows', 'linux'];
-		return Object.fromEntries(
-			Object.entries(builds).sort(([a], [b]) => {
-				const aIndex = osOrder.indexOf(a);
-				const bIndex = osOrder.indexOf(b);
-				const aOrder = aIndex === -1 ? osOrder.length : aIndex;
-				const bOrder = bIndex === -1 ? osOrder.length : bIndex;
-				return aOrder - bOrder;
-			})
-		);
-	});
-
-	function getBuildDisplayName(build: any): string {
-		if (build.os === 'darwin') {
-			return build.arch === 'aarch64' ? 'Apple Silicon' : 'Intel';
-		}
-		if (build.os === 'windows') {
-			return 'MSI';
-		}
-		if (build.os === 'linux') {
-			const arch = build.arch === 'aarch64' ? 'ARM64' : 'Intel';
-			const file: string = build.file?.toLowerCase() ?? '';
-			if (file.includes('appimage')) return `Linux ${arch} (AppImage)`;
-			if (file.includes('.deb')) return `Linux ${arch} (Deb)`;
-			if (file.includes('.rpm')) return `Linux ${arch} (RPM)`;
-			return `Linux ${arch} (CLI)`;
-		}
-		return build.platform;
-	}
 </script>
 
 <div class="release" class:no-separator={!showSeparator}>
@@ -90,30 +47,7 @@
 			{/if}
 
 			{#if downloadLinksVisible}
-				<div class="download-links">
-					{#each Object.entries(groupedBuilds) as [os, builds]}
-						<div class="download-category">
-							<svg
-								class="download-icon"
-								viewBox="0 0 22 22"
-								fill="none"
-								xmlns="http://www.w3.org/2000/svg"
-							>
-								<path
-									d={osIcons[os === 'darwin' ? 'macos' : (os as keyof typeof osIcons)]}
-									fill="currentColor"
-								/>
-							</svg>
-							<div class="download-options">
-								{#each builds as build}
-									<a href={build.url} class="download-link">
-										{getBuildDisplayName(build)}
-									</a>
-								{/each}
-							</div>
-						</div>
-					{/each}
-				</div>
+				<ReleaseDownloadLinks builds={release.builds} />
 			{/if}
 		{/if}
 	</div>
@@ -188,48 +122,6 @@
 
 		&:hover {
 			color: var(--clr-text-1);
-		}
-	}
-
-	.download-links {
-		display: flex;
-		flex-wrap: wrap;
-		padding: 32px;
-		gap: 24px;
-		background-image: radial-gradient(var(--clr-border-2) 1px, transparent 1px);
-		background-size: 6px 6px;
-	}
-
-	.download-category {
-		display: flex;
-		flex-direction: column;
-		min-width: 150px;
-		gap: 12px;
-	}
-
-	.download-icon {
-		width: 22px;
-		height: 22px;
-	}
-
-	.download-options {
-		display: flex;
-		flex-direction: column;
-		gap: 14px;
-	}
-
-	.download-link {
-		width: fit-content;
-		background-color: var(--clr-bg-2);
-		font-size: 14px;
-		font-family: var(--font-mono);
-		text-decoration: underline;
-		text-underline-offset: 2px;
-		transition: all 0.1s ease;
-
-		&:hover {
-			text-decoration: underline wavy;
-			text-decoration-color: var(--clr-theme-pop-element);
 		}
 	}
